chore: add remaining single methods to NWCWalletService class, move shared types out of NWCClient
